Add ib "In Bounds" to refine letters
The function "ib" a.k.a "In Bound" uses a predetermined float to set a range in-which the information is compared to. Basically you call it with a bone's information and give it a target, if the information passed in is within a set amount of target than it returns true.